Menu nav com mudança de cor
