+++ public/app/src/components/Layout/styles.tsx
@@ -14,18 +14,4 @@ export const Container = styled.div`
     border: 3px solid tomato;
   }
 `
-export const Logo = styled.img`
-  width: 180px;
-`
 
-export const MenuList = styled.div`
-  display: flex;
-  flex-direction: column;
-`
-
-export const MenuItems = styled.button`
-  font-size: ${({ theme }) => theme.typography.button};
-  font-weight: bold;
-  cursor: pointer;
-  padding: 1rem;
-`
